+++ src/api/IngresoController.java
@@ -1,6 +1,7 @@
 package api;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
@@ -51,11 +52,11 @@ public class IngresoController extends HttpServlet {
 			st = con.prepareStatement(DatabaseQueries.POST_INGRES);
 			Ingreso c = new Gson().fromJson(request.getReader().readLine(), Ingreso.class);
 			st.setInt(1, c.getUsuarioID());
-			st.setFloat(2, c.getImporte());
+			st.setBigDecimal(2, BigDecimal.valueOf(c.getImporte()));
 			st.setString(3, c.getFecha());
 			
 			st2 = con.prepareStatement(DatabaseQueries.UPDATE_SALARI_USUARIO);
-			st2.setFloat(1, c.getImporte());
+			st2.setBigDecimal(1, BigDecimal.valueOf(c.getImporte()));
 			st2.setInt(2, c.getUsuarioID());
 			
 			if(!st2.execute()) {
